Updated to latest version, requiring modifications to every activity
Added a category for each element on the home page. (Currently not useable, for future use)
This will invalidate the save, but as no one uses this app currently but me and my friend, and he was the one who requested it that's ok.

Updated import to fix an issue caused by the android update to 13, but at least it now has a file picker

fixed a few bugs
added an option to confirm the delete of todos
modified the save button to show the save state in notes

added weekly notifications

Signed-off-by: Peter M <petergmitchell55@gmail.com>